requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==2.2
djangorestframework==3.11.0
idna==2.9
ndg-httpsclient==0.5.1
Pillow==7.0.0
psycopg2==2.8.4
pusher==2.1.4
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
